requirements.txt
kafka-python
redis
psycopg2-binary
fastapi
uvicorn